Rename the ECMWF 46-day archiver
